+++ backend/src/Form/Account/ResetPasswordRequestType.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Account;
 
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\Length;
+use Symfony\Component\Validator\Constraints\Email;
 use Symfony\Component\Validator\Constraints\NotBlank;
 
-class ResetPasswordType extends AbstractType
+class ResetPasswordRequestType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add(child: 'password', type: PasswordType::class, options: [
+            ->add(child: 'email', type: EmailType::class, options: [
+                'label' => null,
                 'attr' => [
-                    'placeholder' => 'Renseignez votre nouveau mot de passe',
+                    'placeholder' => 'Renseignez votre adresse email',
                 ],
                 'constraints' => [
                     new NotBlank([
-                        'message' => 'Vous devez renseigner un mot de passe.',
+                        'message' => 'Vous devez renseigner votre adresse email.',
                     ]),
-                    new Length([
-                        'min' => 8,
-                        'minMessage' => 'Votre mot de passe doit contenir au moins {{ limit }} caractères.',
+                    new Email([
+                        'message' => 'Merci de saisir une adresse email valide.',
                     ]),
                 ],
                 'required' => true,
